backend/foodgram/requirements.txt
asgiref==3.5.2
click==8.1.3
colorama==0.4.4
Django==4.0.5
django-colorfield==0.7.0
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-extra-fields==3.4.0
ez-setup==0.9
flake8==4.0.1
isort==5.10.1
mccabe==0.6.1
Pillow==9.1.1
psycopg2==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.4.0
python-dotenv==0.20.0
pytz==2022.1
reportlab==3.6.10
services==0.1.1
sqlparse==0.4.2
tzdata==2022.1
